fix(administration/dashboards): make standalone
