+++ packages/administration/dashboards-config/components/config-editor.component.ts
