+++ packages/administration/dashboards-config/components/dashboard-edit/dashboard-edit.component.html
