+++ packages/administration/dashboards-config/components/dashboards-list-expand/dashboards-list-expand.component.scss
