+++ packages/administration/dashboards-config/components/dashboards-list/dashboards-list.component.html
@@ -10,7 +10,6 @@
       {{ 'Update indexes' | translate }}
     </button>
 
-
     <button (click)="fileInput.click()"
             [disabled]="loading$ | async"
             mat-button>
