+++ packages/administration/dashboards-config/components/dashboards-list/dashboards-list.component.scss
