+++ packages/administration/dashboards-config/components/schema-editor/schema-editor.component.ts
