+++ packages/administration/dashboards-config/components/selector-text-control/selector-text-control.component.html
