+++ packages/administration/dashboards-config/components/widget-config-examples.component.ts
