+++ packages/administration/dashboards-config/components/widget-copy-dialog/widget-copy-dialog.component.html
