+++ packages/administration/dashboards-config/components/widget-edit/widget-edit.component.html
