+++ packages/administration/dashboards-config/components/widget-edit/widget-list.service.ts
