+++ packages/administration/dashboards-config/index.ts
@@ -1,18 +1,17 @@
-export * from './dashboards.module';
 export * from './dashboards-config.component';
 export * from './const';
 export * from './injectors';
-export * from './widget-edit/widget-edit.component';
-export * from './widget-edit/selector-text-control/selector-text-control.component';
-export * from './widget-edit/schema-editor/schema-editor.component';
-export * from './widget-edit/schema-editor/schema-formly-ext';
-export * from './widget-edit/widget-config-examples.component';
-export * from './widget-edit/widget-list.service';
-export * from './dashboard-edit/dashboard-edit.component';
-export * from './dashboard-editor.service';
-export * from './dashboards-list/dashboards-list-expand/dashboards-list-expand.component';
-export * from './dashboards-list/dashboards-export.service';
-export * from './dashboards-list/dashboards-import.service';
-export * from './dashboards-list/dashboards-list.component';
-export * from './dashboards-manager.service';
+export * from './components/widget-edit/widget-edit.component';
+export * from './components/selector-text-control/selector-text-control.component';
+export * from './components/schema-editor/schema-editor.component';
+export * from './components/schema-editor/schema-formly-ext';
+export * from './components/widget-config-examples.component';
+export * from './components/widget-edit/widget-list.service';
+export * from './components/dashboard-edit/dashboard-edit.component';
+export * from './services/dashboard-editor.service';
+export * from './components/dashboards-list-expand/dashboards-list-expand.component';
+export * from './components/dashboards-list/dashboards-export.service';
+export * from './components/dashboards-list/dashboards-import.service';
+export * from './components/dashboards-list/dashboards-list.component';
+export * from './services/dashboards-manager.service';
 export * from './xm-dynamic-list.module';
