+++ packages/administration/dashboards-config/services/dashboards-manager.service.ts
