+++ packages/administration/dashboards-config/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './mock-widget-list.service';
