+++ packages/administration/dashboards-config/testing/mock-widget-list.service.ts
