+++ packages/administration/route/admin.module.ts
@@ -3,7 +3,7 @@ import { RouterModule } from '@angular/router';
 import { AuditsModule } from '@xm-ngx/administration/audits';
 import { RolesManagementDetailModule } from '@xm-ngx/administration/roles-management-detail';
 import { UserManagementDetailModule } from '@xm-ngx/administration/user-management-detail';
-import { adminState } from './admin.route';
+import { ADMIN_ROUTES } from './admin.route';
 import { ClientManagementModule } from '@xm-ngx/administration/client-management';
 import { ClientResolvePagingParams } from './client-management.route';
 import { DocsModule } from '@xm-ngx/administration/docs';
@@ -36,7 +36,7 @@ import { UserResolvePagingParams } from './user-management.route';
         TranslationModule,
         UserManagementModule,
         UserManagementDetailModule,
-        RouterModule.forChild(adminState),
+        RouterModule.forChild(ADMIN_ROUTES),
     ],
     declarations: [],
     providers: [
