+++ packages/administration/route/admin.route.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 
 import { UserRouteAccessService } from '@xm-ngx/core/permission';
-import { adminConfigState } from './admin-config.route';
+import { ADMIN_CONFIG_ROUTE } from './admin-config.route';
 import { auditsRoute } from './audits.route';
 import { clientMgmtRoute } from './client-management.route';
 import { docsRoute } from './docs.route';
@@ -15,32 +15,32 @@ import { rolesMgmtRoute } from './roles-management.route';
 import { rolesMatrixRoute } from './roles-matrix.route';
 import { translationRoute } from './translation.route';
 import { userMgmtRoute } from './user-management.route';
+import { DASHBOARDS_ROUTE } from './dashboards.route';
 
-
-const ADMIN_ROUTES = [
-    auditsRoute,
-    docsRoute,
-    healthRoute,
-    formPlaygroundRoute,
-    maintenanceRoute,
-    translationRoute,
-    logsRoute,
-    gatewayRoute,
-    rolesMatrixRoute,
-    ...userMgmtRoute,
-    ...clientMgmtRoute,
-    ...rolesMgmtRoute,
-    metricsRoute,
-    ...adminConfigState,
-];
-
-export const adminState: Routes = [{
-    path: '',
-    data: {
-        authorities: ['ROLE_ADMIN'],
+export const ADMIN_ROUTES: Routes = [
+    {
+        path: '',
+        data: {
+            authorities: ['ROLE_ADMIN'],
+        },
+        // canActivate: [UserRouteAccessService],
+        canActivateChild: [UserRouteAccessService],
+        children: [
+            auditsRoute,
+            docsRoute,
+            healthRoute,
+            formPlaygroundRoute,
+            maintenanceRoute,
+            translationRoute,
+            logsRoute,
+            gatewayRoute,
+            rolesMatrixRoute,
+            ...userMgmtRoute,
+            ...clientMgmtRoute,
+            ...rolesMgmtRoute,
+            metricsRoute,
+            ADMIN_CONFIG_ROUTE,
+            DASHBOARDS_ROUTE
+        ]
     },
-    // canActivate: [UserRouteAccessService],
-    canActivateChild: [UserRouteAccessService],
-    children: ADMIN_ROUTES,
-},
 ];
